+++ encoding/codecv0/codecv0_test.go
@@ -7,9 +7,10 @@ import (
 	"os"
 	"testing"
 
+	"github.com/stretchr/testify/assert"
+
 	"github.com/scroll-tech/go-ethereum/common"
 	"github.com/scroll-tech/go-ethereum/log"
-	"github.com/stretchr/testify/assert"
 
 	"github.com/scroll-tech/da-codec/encoding"
 )
