+++ encoding/codecv4/codecv4.go
@@ -15,6 +15,7 @@ import (
 	"github.com/scroll-tech/go-ethereum/log"
 
 	"github.com/scroll-tech/da-codec/encoding"
+	"github.com/scroll-tech/da-codec/encoding/codecv1"
 	"github.com/scroll-tech/da-codec/encoding/codecv3"
 	"github.com/scroll-tech/da-codec/encoding/zstd"
 )
@@ -27,6 +28,9 @@ type DABlock = codecv3.DABlock
 
 // DAChunk groups consecutive DABlocks with their transactions.
 type DAChunk = codecv3.DAChunk
+
+// DAChunkRawTx groups consecutive DABlocks with their L2 transactions, L1 msgs are loaded in another place.
+type DAChunkRawTx = codecv3.DAChunkRawTx
 
 // DABatch contains metadata about a batch of DAChunks.
 type DABatch struct {
@@ -57,6 +61,11 @@ func NewDABlock(block *encoding.Block, totalL1MessagePoppedBefore uint64) (*DABl
 // NewDAChunk creates a new DAChunk from the given encoding.Chunk and the total number of L1 messages popped before.
 func NewDAChunk(chunk *encoding.Chunk, totalL1MessagePoppedBefore uint64) (*DAChunk, error) {
 	return codecv3.NewDAChunk(chunk, totalL1MessagePoppedBefore)
+}
+
+// DecodeDAChunksRawTx takes a byte slice and decodes it into a []*DAChunkRawTx.
+func DecodeDAChunksRawTx(bytes [][]byte) ([]*DAChunkRawTx, error) {
+	return codecv3.DecodeDAChunksRawTx(bytes)
 }
 
 // NewDABatch creates a DABatch from the provided encoding.Batch.
@@ -237,6 +246,23 @@ func ConstructBlobPayload(chunks []*encoding.Chunk, enableCompress bool, useMock
 	copy(z[start:], pointBytes)
 
 	return blob, blobVersionedHash, &z, blobBytes, nil
+}
+
+// DecodeTxsFromBlob decodes txs from blob bytes and writes to chunks
+func DecodeTxsFromBlob(blob *kzg4844.Blob, chunks []*DAChunkRawTx) error {
+	rawBytes := encoding.BytesFromBlobCanonical(blob)
+
+	// if first byte is 1 - data compressed, 0 - not compressed
+	if rawBytes[0] == 0x1 {
+		magics := []byte{0x28, 0xb5, 0x2f, 0xfd}
+		blobBytes, err := encoding.DecompressScrollBatchBytes(append(magics, rawBytes[1:]...))
+		if err != nil {
+			return err
+		}
+		return codecv1.DecodeTxsFromBytes(blobBytes, chunks, MaxNumChunks)
+	} else {
+		return codecv1.DecodeTxsFromBytes(rawBytes[1:], chunks, MaxNumChunks)
+	}
 }
 
 // NewDABatchFromBytes decodes the given byte slice into a DABatch.
